seccion_b_etl/requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.4
SQLAlchemy==2.0.23